+++ backend/tests/baserow/api/user_files/test_user_files_views.py
@@ -144,7 +144,14 @@ def test_upload_file(api_client, data_fixture, tmpdir):
 
 @pytest.mark.django_db
 @responses.activate
-def test_upload_file_via_url(api_client, data_fixture, tmpdir):
+@patch('baserow.core.user_files.handler.socket.gethostbyname')
+def test_upload_file_via_url(mock_gethostbyname, api_client, data_fixture, tmpdir):
+    # A public ip address on the internet and not a private network. By mocking this
+    # return value, the is_private check will always return False and will therefore
+    # never raise the InvalidFileURLError because the file does not needs to be fetched
+    # from a private network.
+    mock_gethostbyname.return_value = '171.171.171.171'
+
     user, token = data_fixture.create_user_and_token(
         email='test@test.nl', password='password', first_name='Test1'
     )
@@ -172,6 +179,14 @@ def test_upload_file_via_url(api_client, data_fixture, tmpdir):
     )
     assert response.status_code == HTTP_400_BAD_REQUEST
     assert response.json()['error'] == 'ERROR_FILE_URL_COULD_NOT_BE_REACHED'
+
+    response = api_client.post(
+        reverse('api:user_files:upload_via_url'),
+        data={'url': 'ftp://localhost/test2.txt'},
+        HTTP_AUTHORIZATION=f'JWT {token}'
+    )
+    assert response.status_code == HTTP_400_BAD_REQUEST
+    assert response.json()['error'] == 'ERROR_INVALID_FILE_URL'
 
     responses.add(
         responses.GET,