+++ changelog.md
@@ -10,6 +10,7 @@
 * Show the number of filters and sorts active in the header of a grid view.
 * The first user to sign-up after installation now gets given staff status.
 * Prevented the date field value to be negative.
+* Fixed SSRF bug in the file upload by URL by blocking urls to the private network.
 
 ## Released (2021-03-01)